pyproject.toml
[project]
name = "ai-course-architect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "arxiv>=2.4.0",
    "ddgs>=9.10.0",
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "neo4j>=6.1.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "wikipedia>=1.4.0",
    "wikipedia-api>=0.9.0",
]
